+++ contracts/utils/FaucetToken.sol
@@ -7,23 +7,19 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract FaucetToken is ERC20 {
     uint8 private _decimals;
 
-    constructor(
-        string memory _name,
-        string memory _symbol,
-        uint8 __decimals
-    ) ERC20(_name, _symbol) {
+    constructor(string memory _name, string memory _symbol, uint8 __decimals) ERC20(_name, _symbol) {
         _decimals = __decimals;
     }
-    
+
     function decimals() public view virtual override returns (uint8) {
         return _decimals;
     }
 
-    function mint(address _to, uint256 _amount) external {
+    function mint(address _to, uint _amount) external {
         _mint(_to, _amount);
     }
 
-    function burn(uint256 _amount) external {
+    function burn(uint _amount) external {
         _burn(msg.sender, _amount);
     }
 }
